+++ MEM-Win32App-Builder/MEM-Win32App-Builder.psm1
@@ -3,7 +3,7 @@ $public  = @(Get-ChildItem -Path (Join-Path -Path $PSScriptRoot -ChildPath 'Publ
 $private = @(Get-ChildItem -Path (Join-Path -Path $PSScriptRoot -ChildPath 'Private/*.ps1') -Recurse -ErrorAction Stop)
 foreach ($import in @($public + $private)) {
     try {
-        Write-Host "Importing $import.FullName"
+        # Write-Host "Importing $import.FullName"
         . $import.FullName
     }
     catch {