+++ MEM-Win32App-Builder/template/detection/detection.ps1
@@ -1,0 +1,12 @@
+<#
+.SYNOPSIS
+
+.DESCRIPTION
+
+.EXAMPLE
+
+.NOTES
+    Author:
+    Date:
+    Version:
+#>