+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.1] Unreleased
+
+* Changed private function `downloadWin32AppUtil.ps1` to public function `Download-Win32AppUtil.ps1`
+* Updated template files
+* Added docs
+* Removed Ubuntu and macOS from GitHub Actions CI
+
 ## [0.1.0] Unreleased
 
 * Created the module using [Stucco](https://github.com/devblackops/Stucco)
+* Created public function `New-MemWin32AppBuild.ps1`
+* Created private function `downloadWin32AppUtil.ps1`
+* Created Windows Sandbox config for testing